250410 lecture update

+++ lecture/02-02-many-to-one-relationships/articles/views.py
@@ -1,10 +1,12 @@
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import redirect, render
+from django.views.decorators.http import require_http_methods, require_safe, require_POST
 
 from .forms import ArticleForm, CommentForm
 from .models import Article, Comment
 
 
+@require_safe
 def index(request):
     articles = Article.objects.all()
     context = {
@@ -13,6 +15,7 @@ def index(request):
     return render(request, 'articles/index.html', context)
 
 
+@require_safe
 def detail(request, pk):
     article = Article.objects.get(pk=pk)
     comment_form = CommentForm()
@@ -27,11 +30,16 @@ def detail(request, pk):
 
 
 @login_required
+@require_http_methods(['GET', 'POST'])
 def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            # 현재 유저가 누구인지 DB에서 찾아야 하는가?
+            # 이미 요청 객체에 어떤 사람인지 들어있기 때문에 조회하지 않는다.
+            article.user = request.user
+            article.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -42,22 +50,29 @@ def create(request):
 
 
 @login_required
+@require_POST
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
+    if request.user == article.user:
+        article.delete()
     return redirect('articles:index')
 
 
 @login_required
+@require_http_methods(['GET', 'POST'])
 def update(request, pk):
     article = Article.objects.get(pk=pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article.pk)
+    # 수정을 요청시도하는 사용자와 게시글의 작성자가 같은지를 확인
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:index')
     context = {
         'article': article,
         'form': form,
@@ -65,6 +80,8 @@ def update(request, pk):
     return render(request, 'articles/update.html', context)
 
 
+@login_required
+@require_POST
 def comments_create(request, article_pk):
     # 어떤 게시글에 작성되는 댓글이 알려면 게시글을 먼저 조회
     article = Article.objects.get(pk=article_pk)
@@ -77,6 +94,7 @@ def comments_create(request, article_pk):
         # 댓글 인스턴스는 생성해주지만 실제 DB에 아직 저장 요청은 보내지 않고 대기
         comment = comment_form.save(commit=False)
         comment.article = article
+        comment.user = request.user
         comment.save()
         return redirect('articles:detail', article.pk)
     context = {
@@ -86,13 +104,16 @@ def comments_create(request, article_pk):
 
 
 # 게시글 pk를 가져오는 두번째 방법 (url에서 넘겨주기)
+@login_required
+@require_POST
 def comments_delete(request, article_pk, comment_pk):
     # 어떤 댓글이 삭제되는 것인지 조회
     comment = Comment.objects.get(pk=comment_pk)
     # 게시글 pk를 가져오는 첫번째 방법 (댓글 삭제 전에 게시글 번호 저장해두기)
     # article_id = comment.article.pk
     # 댓글 삭제
-    comment.delete()
+    if request.user == comment.user:
+        comment.delete()
     # 첫번째 방법에 대한 return
     # return redirect('articles:detail', article_id)
     # 두번째 방법에 대한 return
